refactor: Refactoring to integrate  NPC-registry.
