+++ adventures.and.beyond/src/entity/manager/EntityManagerImpl.java
@@ -1,28 +1,28 @@
 package entity.manager;
 
-import dialogue.manager.DialogueKey;
-import directionEnum.Direction;
-
 import entity.npc.DrawableNonPlayerCharacter;
-import entity.npc.NonPlayerCharacterFactory;
-import entity.npc.NonPlayerCharacterFactoryImpl;
 import entity.npc.UpdatableNonPlayerCharacter;
+import entity.npc.registry.NonPlayerCharacterRegistry;
+import entity.npc.registry.NonPlayerCharacterRegistryImpl;
 import entity.player.Player;
 
 import java.awt.*;
-import java.util.ArrayList;
 import java.util.List;
 
 import static helper.Constant.TILE_SIZE;
 
 public class EntityManagerImpl implements EntityManager {
+
     private Player mainEntity;
 
-    private List<UpdatableNonPlayerCharacter> secondaryUpdatables;
-    private List<DrawableNonPlayerCharacter> secondaryDrawables;
+    private List<UpdatableNonPlayerCharacter> secondaryUpdatableList;
+    private List<DrawableNonPlayerCharacter> secondaryDrawableList;
+
+    private NonPlayerCharacterRegistry nonPlayerCharacterRegistry;
 
 
     public EntityManagerImpl() {
+        this.nonPlayerCharacterRegistry = new NonPlayerCharacterRegistryImpl();
         this.registerEntities();
     }
 
@@ -36,22 +36,14 @@ public class EntityManagerImpl implements EntityManager {
     }
 
     private void registerSecondaryEntities() {
-        NonPlayerCharacterFactory nonPlayerCharacterFactory = new NonPlayerCharacterFactoryImpl();
-        String[] characterTypes = nonPlayerCharacterFactory.getCharacterTypes();
-
-        nonPlayerCharacterFactory.register(characterTypes[0], 28 * TILE_SIZE, 12 * TILE_SIZE, Direction.FACING_FORWARD, false, DialogueKey.DIALOGUE_01);
-        nonPlayerCharacterFactory.register(characterTypes[0], 37 * TILE_SIZE, 12 * TILE_SIZE, Direction.FACING_FORWARD, false, DialogueKey.DIALOGUE_02);
-        nonPlayerCharacterFactory.register(characterTypes[0], 23 * TILE_SIZE, 12 * TILE_SIZE, Direction.FACING_FORWARD, false, DialogueKey.DIALOGUE_01);
-
-
-        this.secondaryUpdatables = nonPlayerCharacterFactory.getUpdatableNonPlayerCharacterList();
-        this.secondaryDrawables = nonPlayerCharacterFactory.getDrawableNonPlayerCharacterList();
+        this.secondaryUpdatableList = nonPlayerCharacterRegistry.getRegisteredUpdatableList();
+        this.secondaryDrawableList = nonPlayerCharacterRegistry.getRegisteredDrawableList();
     }
 
     public void drawEntities(Graphics2D graphics2D) {
         this.mainEntity.draw(graphics2D);
 
-        for (DrawableNonPlayerCharacter character : this.secondaryDrawables) {
+        for (DrawableNonPlayerCharacter character : this.secondaryDrawableList) {
             character.draw(graphics2D, this.getPlayer());
         }
     }
@@ -59,7 +51,7 @@ public class EntityManagerImpl implements EntityManager {
     public void updateEntities() {
         this.mainEntity.update();
 
-        for (UpdatableNonPlayerCharacter character : this.secondaryUpdatables) {
+        for (UpdatableNonPlayerCharacter character : this.secondaryUpdatableList) {
             character.update();
         }
     }
